Refactor page structure and add new sections for About, Flow, Partners, Schedule, Speakers, Venue, and Footer
